Fixed verification failure bug for long arrays
There was a bug in reference data generation.

The numElements has to be a multiple of 256 else the issue is still
observed in optimized kernel. It is because the workitems in the last
work-group do not read the final values correctly.

+++ firFilter.cl
@@ -47,7 +47,7 @@ void NaiveMovingAverageFilter(
 
     size_t halfFilterLength = filterLength / 2;
 
-    size_t windowIdx = idx - halfFilterLength;
+    long windowIdx = idx - halfFilterLength;
     float  value_r = 0.0f;
 	float  value_i = 0.0f;
 	float fw = 0.0f;
