+++ firFilter.cpp
@@ -31,16 +31,16 @@ void GenerateTestData(size_t const numElements,
     }
 
 	if (verify) {
-		size_t const filterLength = filterWeights.size();
-		size_t const halfFilterLength = filterLength / 2;
+		long const filterLength = filterWeights.size();
+		long const halfFilterLength = filterLength / 2;
 
 		for (size_t i = 0; i < numElements; i++)
 		{
-			size_t windowIdx = i - halfFilterLength;
+			long windowIdx = i - halfFilterLength;
 			float  value_r = 0.0f;
 			float  value_i = 0.0f;
 
-			for (size_t filterIdx = 0; filterIdx < filterLength; filterIdx++, windowIdx++)
+			for (int filterIdx = 0; filterIdx < filterLength; filterIdx++, windowIdx++)
 			{
 				float sample_r = (0 <= windowIdx && windowIdx < numElements) ? testData_r[windowIdx] : 0.0f;
 				value_r += sample_r * filterWeights[filterIdx];
